refactor: 메시지 상수로 분리
1. ErrorMessage 신규 생성 및 메시지 상수로 분리
2. exception throw 시 errorMessage 상수 사용하도록 수정
3. Test 케이스 내에 ERROR MESSAGE 상수 사용하도록 수정
