+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.service.RacingCarsGame;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
