+++ src/main/java/racingcar/Constants/ErrorMessage.java
@@ -1,0 +1,10 @@
+package racingcar.Constants;
+
+public class ErrorMessage {
+
+    public static final String ERR_MSG_TITLE = "[ERROR]";
+    public static final String CAR_NAME_NOT_VALID_ERR_MSG = ERR_MSG_TITLE + " 자동차 이름은 1~4자";
+    public static final String CAR_NAME_DUPLICATED_ERR_MSG = ERR_MSG_TITLE + " 자동차 이름 중복";
+    public static final String MOVE_COUNT_VALID_ERR_MSG = ERR_MSG_TITLE + " 시도할 횟수는 0이상 숫자만 입력 가능";
+
+}
