+++ src/main/java/racingcar/domain/Cars.java
@@ -1,11 +1,13 @@
-package racingcar;
+package racingcar.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import view.UserInterFace;
+import racingcar.view.UserInterFace;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+
+import static racingcar.Constants.ErrorMessage.CAR_NAME_DUPLICATED_ERR_MSG;
 
 public class Cars {
 
@@ -22,7 +24,7 @@ public class Cars {
             carNames.add(car.toString());
         }
         if(carNames.size() != cars.size()){
-            throw new IllegalArgumentException("[ERROR] 자동차 이름 중복");
+            throw new IllegalArgumentException(CAR_NAME_DUPLICATED_ERR_MSG);
         }
     }
