+++ src/main/java/racingcar/view/UserInterFace.java
@@ -1,4 +1,4 @@
-package view;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
